refactor pulumi code to get ip addr
